modified:   hopalong_advanced.py

+++ hopalong_advanced.py
@@ -59,7 +59,7 @@ def compute_trajectory(a, b, c, num):
 def generate_trajectory_image(points, image_size):
     # Generates an image array with the mapped trajectory points
     img_width, img_height = image_size
-    image = np.zeros((img_height, img_width), dtype=np.uint32)
+    image = np.zeros((img_height, img_width), dtype=np.uint16)
 
     min_x, max_x = np.min(points[:, 0]), np.max(points[:, 0])
     min_y, max_y = np.min(points[:, 1]), np.max(points[:, 1])
